refactor: StagesAndEventsWidget to WidgetStagesAndEvents
