+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/Stage/Stage.component.js
