+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/Stage/StageOverview/index.js
