+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/Stage/StageOverview/stageOverview.types.js
@@ -1,5 +1,5 @@
 // @flow
-import type { Icon } from '../../../../../../metaData';
+import type { Icon } from '../../../../../metaData';
 
 export type Props = {|
     title: string,
