+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/Stage/index.js
