+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/Stages.component.js
