+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/index.js
