+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/WidgetStagesAndEvents.component.js
@@ -1,11 +1,11 @@
 // @flow
 import React, { useState, useCallback } from 'react';
 import i18n from '@dhis2/d2-i18n';
-import { Widget } from '../../Widget';
+import { Widget } from '../Widget';
 import { Stages } from './Stages';
 import type { Props } from './stagesAndEvents.types';
 
-export const StagesAndEventsWidget = ({ className, stages }: Props) => {
+export const WidgetStagesAndEvents = ({ className, stages }: Props) => {
     const [open, setOpenStatus] = useState(true);
 
     return (
