+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { WidgetStagesAndEvents } from './WidgetStagesAndEvents.component';
