+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/stagesAndEvents.types.js
@@ -1,5 +1,5 @@
 // @flow
-import type { ProgramStage } from '../../../metaData';
+import type { ProgramStage } from '../../metaData';
 
 export type Props = {|
     stages: Map<string, ProgramStage>,
