--- src/core_modules/capture-core/components/Widgets/StagesAndEventsWidget/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { StagesAndEventsWidget } from './StagesAndEventsWidget.component';
